remove builder pattern and fix review changes

+++ core/org.wso2.carbon.nextgen.config/src/main/java/org/wso2/carbon/nextgen/config/MetaDataParser.java
@@ -41,11 +41,13 @@ import java.util.Set;
 /**
  * ConfigurationFileMetadata Persisting parser.
  */
-public class MetaDataParser {
+class MetaDataParser {
 
-    private static Log logger = LogFactory.getLog(MetaDataParser.class);
+    private static Log log = LogFactory.getLog(MetaDataParser.class);
 
-    public static Map<String, String> readLastModifiedValues(String basePath, String path)
+    private MetaDataParser() {}
+
+    static Map<String, String> readLastModifiedValues(String basePath, String path)
             throws ConfigParserException {
 
         Map<String, String> md5sumValues = new HashMap<>();
@@ -64,7 +66,7 @@ public class MetaDataParser {
         return md5sumValues;
     }
 
-    public static String readLastModifiedValue(String path)
+    private static String readLastModifiedValue(String path)
             throws ConfigParserException {
 
         try {
@@ -97,8 +99,8 @@ public class MetaDataParser {
         }
     }
 
-    public static ChangedFileSet getChangedFiles(String basePath, List<String> deploymentConfigurationPaths,
-                                                 String metadataFilePath)
+    static ChangedFileSet getChangedFiles(String basePath, List<String> deploymentConfigurationPaths,
+                                          String metadataFilePath)
             throws ConfigParserException {
 
         File metaDataFile = new File(metadataFilePath);
@@ -110,7 +112,6 @@ public class MetaDataParser {
         try (FileInputStream fileInputStream = new FileInputStream(metaDataFile)) {
             properties.load(fileInputStream);
         } catch (IOException e) {
-            logger.error("Metadata File couldn't Read", e);
             throw new ConfigParserException("Metadata File couldn't Read", e);
         }
         ChangedFileSet changedFileSet = new ChangedFileSet();
@@ -132,29 +133,29 @@ public class MetaDataParser {
         return changedFileSet;
     }
 
-    public static ChangedFileSet getChangedFiles(String basePath,
-                                                 String metadataFilePath)
+    static ChangedFileSet getChangedFiles(String basePath,
+                                          String metadataFilePath)
             throws ConfigParserException {
 
         File metaDataFile = new File(metadataFilePath);
         if (!metaDataFile.exists()) {
-            return new ChangedFileSet(true, Collections.EMPTY_LIST, Collections.EMPTY_LIST);
+            return new ChangedFileSet(true, Collections.emptyList(), Collections.emptyList());
         }
         Properties properties = new Properties();
 
         try (FileInputStream fileInputStream = new FileInputStream(metaDataFile)) {
             properties.load(fileInputStream);
         } catch (IOException e) {
-            logger.error("Metadata File couldn't Read", e);
+            log.error("Metadata File couldn't Read", e);
             throw new ConfigParserException("Metadata File couldn't Read", e);
         }
         ChangedFileSet changedFileSet = new ChangedFileSet();
         for (Map.Entry<Object, Object> entry : properties.entrySet()) {
             String path = (String) entry.getKey();
-            String lastmodifiedValue = (String) entry.getValue();
+            String lastModifiedValue = (String) entry.getValue();
             String actualLastModifiedValue = readLastModifiedValue(Paths.get(basePath, path).toString());
             if (StringUtils.isNotEmpty(actualLastModifiedValue)) {
-                if (!lastmodifiedValue.equals(actualLastModifiedValue)) {
+                if (!lastModifiedValue.equals(actualLastModifiedValue)) {
                     changedFileSet.addChangedFile(path);
                 }
             }
@@ -162,13 +163,13 @@ public class MetaDataParser {
         return changedFileSet;
     }
 
-    @edu.umd.cs.findbugs.annotations.SuppressFBWarnings(value = "RV_RETURN_VALUE_IGNORED_BAD_PRACTICE",
-            justification = "return not need in mkdirs()")
-    public static void storeMetaDataEntries(String basePath, String outputFilePath, Set<String> entries)
+    static void storeMetaDataEntries(String basePath, String outputFilePath, Set<String> entries)
             throws ConfigParserException {
 
         File outputFile = new File(outputFilePath);
-        outputFile.getParentFile().mkdirs();
+        if (!outputFile.getParentFile().exists() && !outputFile.getParentFile().mkdirs()) {
+            throw new ConfigParserException("Error while creating new directory " + outputFile.getAbsolutePath());
+        }
         Properties properties = new Properties();
         for (String entry : entries) {
             properties.putAll(readLastModifiedValues(basePath, entry));
@@ -176,18 +177,17 @@ public class MetaDataParser {
         try (FileOutputStream fileOutputStream = new FileOutputStream(outputFilePath)) {
             properties.store(fileOutputStream, null);
         } catch (IOException e) {
-            logger.error("error while storing metadata", e);
-            throw new RuntimeException("error while storing metadata");
+            throw new ConfigParserException("Error while storing metadata", e);
         }
     }
 
-    @edu.umd.cs.findbugs.annotations.SuppressFBWarnings(value = "RV_RETURN_VALUE_IGNORED_BAD_PRACTICE",
-            justification = "return not need in mkdirs()")
-    public static void storeReferences(String metadataPropertyPath, Context context)
+    static void storeReferences(String metadataPropertyPath, Context context)
             throws ConfigParserException {
 
         File outputFile = new File(metadataPropertyPath);
-        outputFile.getParentFile().mkdirs();
+        if (!outputFile.getParentFile().exists() && !outputFile.getParentFile().mkdirs()) {
+            throw new ConfigParserException("Error while creating new directory " + outputFile.getAbsolutePath());
+        }
         try (OutputStreamWriter outputStreamWriter = new OutputStreamWriter(new FileOutputStream(metadataPropertyPath),
                 StandardCharsets.UTF_8)) {
             Properties properties = new Properties();
@@ -203,7 +203,7 @@ public class MetaDataParser {
         }
     }
 
-    public static boolean isReferencesChanged(String metadataPropertyPath) {
+    static boolean isReferencesChanged(String metadataPropertyPath) {
 
         boolean status = false;
         if (new File(metadataPropertyPath).exists()) {
@@ -211,7 +211,7 @@ public class MetaDataParser {
             try (FileInputStream fileInputStream = new FileInputStream(metadataPropertyPath)) {
                 references.load(fileInputStream);
             } catch (IOException e) {
-                logger.error("Error while reading References", e);
+                log.error("Error while reading References", e);
             }
             for (Map.Entry<Object, Object> entry : references.entrySet()) {
                 String key = (String) entry.getKey();
